+++ lib/pgchief/command/database_restore.rb
@@ -10,11 +10,14 @@ module Pgchief
 
       attr_reader :database, :filename
 
+      def_delegators :s3, :configured?, :client, :bucket, :path
+
       def call
         @database = params.first
         @filename = params.last
         raise Pgchief::Errors::DatabaseMissingError unless db_exists?
 
+        download! if configured?
         restore!
 
         "Database '#{database}' restored from #{filename}"
@@ -25,6 +28,14 @@ module Pgchief
       end
 
       private
+
+      def download!
+        client.get_object(
+          bucket: bucket,
+          key: "#{path}#{filename}",
+          response_target: local_location
+        )
+      end
 
       def restore!
         `pg_restore --clean --no-owner --dbname=#{Pgchief::Config.pgurl}/#{database} #{local_location}`
@@ -37,6 +48,10 @@ module Pgchief
 
       def local_location
         "#{Pgchief::Config.backup_dir}/#{filename}"
+      end
+
+      def s3
+        Pgchief::Config.s3
       end
     end
   end